# docs/specs/api.openapi.yaml
openapi: 3.1.0
info:
  title: VedioGen Control API
  version: 0.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    P0 创作主链路和内部模型管理后台的控制面 API。
    长任务返回 Run ID，通过事件流或查询接口获取状态。
servers:
  - url: /api/v1
...
paths:
...
  /account-brief:
    get:
      summary: Get the default creator account brief
      tags: [Account Brief]
      operationId: getDefaultAccountBrief
      security:
        - userSession: []
      responses:
        "200":
          description: Current account brief
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountBrief"
        default:
          $ref: "#/components/responses/Error"
    put:
      summary: Create or replace the default creator account brief
      tags: [Account Brief]
      operationId: upsertDefaultAccountBrief
      security:
        - userSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountBriefInput"
      responses:
        "200":
          description: Account brief saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountBrief"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
  /projects:
    get:
      summary: List projects
      tags: [Projects]
      operationId: listProjects
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProjectStatus"
        - name: contentPackId
          in: query
          schema:
            type: string
        - name: query
          in: query
          description: Case-insensitive project title search.
          schema:
            type: string
            maxLength: 120
      responses:
        "200":
          description: Project page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectPage"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: Create a project from an initial idea
      tags: [Projects]
      operationId: createProject
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
  /projects/{projectId}:
    get:
      summary: Get a project
      tags: [Projects]
      operationId: getProject
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
    patch:
      summary: Update editable project metadata
      tags: [Projects]
      operationId: updateProject
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 120
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
  /projects/{projectId}/workspace:
    get:
      summary: Get a restorable project workspace snapshot
      tags: [Projects]
      operationId: getProjectWorkspace
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project, conversation, facts, evidence, assets, and current run references
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectWorkspace"
        default:
          $ref: "#/components/responses/Error"
  /projects/{projectId}/messages:
    post:
      summary: Add a user message to a project
      tags: [Projects, Advisor]
      operationId: addProjectMessage
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "201":
          description: Message accepted and facts extraction scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationMessage"
        default:
          $ref: "#/components/responses/Error"
  /projects/{projectId}/events:
    get:
      summary: Stream project and run events
      tags: [Projects]
      operationId: streamProjectEvents
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Server-sent project and run events
          content:
            text/event-stream:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Error"
  /projects/{projectId}/advisor-runs:
    post:
      summary: Start or revise Creative Advisor proposals
      tags: [Advisor]
      operationId: createAdvisorRun
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                instruction:
                  type: string
                  maxLength: 1000
                sourceProposalKeys:
                  type: array
                  maxItems: 3
                  uniqueItems: true
                  items:
                    type: string
      responses:
        "202":
          description: Advisor run queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdvisorRun"
        default:
          $ref: "#/components/responses/Error"
  /advisor-runs/{advisorRunId}:
    get:
      summary: Get an Advisor run and its proposals
      tags: [Advisor]
      operationId: getAdvisorRun
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/AdvisorRunId"
      responses:
        "200":
          description: Advisor run and proposal result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdvisorRunDetail"
        default:
          $ref: "#/components/responses/Error"
  /projects/{projectId}/creative-briefs:
    post:
      summary: Create a Creative Brief draft
      tags: [Briefs]
      operationId: createCreativeBriefDraft
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./creative-brief.schema.json"
      responses:
        "201":
          description: Brief draft created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionedDocument"
        default:
          $ref: "#/components/responses/Error"
  /projects/{projectId}/creative-briefs/{briefVersionId}:
    get:
      summary: Get a Creative Brief version
      tags: [Briefs]
      operationId: getCreativeBriefVersion
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/BriefVersionId"
      responses:
        "200":
          description: Creative Brief version
          content:
            application/json:
              schema:
                $ref: "./creative-brief.schema.json"
        default:
          $ref: "#/components/responses/Error"
  /projects/{projectId}/creative-briefs/{briefVersionId}/approval:
    post:
      summary: Approve an immutable Creative Brief version
      tags: [Briefs]
      operationId: approveCreativeBrief
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/BriefVersionId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalRequest"
      responses:
        "200":
          description: Brief approved and downstream stale markers updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
  /projects/{projectId}/storyboard-runs:
    post:
      summary: Generate a Storyboard draft
      tags: [Storyboards]
      operationId: generateStoryboard
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [creativeBriefVersionId]
              properties:
                creativeBriefVersionId:
                  $ref: "#/components/schemas/Uuid"
                instruction:
                  type: string
                  maxLength: 1000
      responses:
        "202":
          description: Storyboard generation queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncRun"
        default:
          $ref: "#/components/responses/Error"
  /projects/{projectId}/storyboards/{storyboardVersionId}:
    get:
      summary: Get a Storyboard version
      tags: [Storyboards]
      operationId: getStoryboard
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/StoryboardVersionId"
      responses:
        "200":
          description: Storyboard version
          content:
            application/json:
              schema:
                $ref: "./storyboard.schema.json"
        default:
          $ref: "#/components/responses/Error"
    put:
      summary: Replace an editable Storyboard draft
      tags: [Storyboards]
      operationId: replaceStoryboardDraft
      security:
        - userSession: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/StoryboardVersionId"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./storyboard.schema.json"
      responses:
        "200":
          description: Draft replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionedDocument"
        "409":
          $ref: "#/components/responses/Conflict"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    userSession:
      type: http
      scheme: bearer
      bearerFormat: session
      description: P0 local mode injects a fixed local user; external deployment requires authentication.
    adminSession:
      type: http
      scheme: bearer
      bearerFormat: session
      description: Requires model-admin; P0 may restrict the admin route to localhost.
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
...
  responses:
    Error:
      description: Error response
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: Version, state, or idempotency conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    AccountBrief:
      type: object
      additionalProperties: false
      required: [id, contentPackId, accountGoal, audience, tone, publishingCadence, preferences, version, rowVersion, createdAt, updatedAt]
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        contentPackId:
          type: string
        accountGoal:
          type: string
        audience:
          type: array
          items:
            type: string
        tone:
          type: array
          items:
            type: string
        publishingCadence:
          enum: [daily, weekdays, weekly, custom]
        preferences:
          type: object
          additionalProperties: true
        version:
          type: integer
          minimum: 1
        rowVersion:
          type: integer
          minimum: 1
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        updatedAt:
          $ref: "#/components/schemas/Timestamp"
...